# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing medicine coupons and discounts",
        "title": "Farmako Coupon API",
        "contact": {
            "name": "API Support",
            "email": "support@pharmaapp.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/admin/addCoupons": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a new coupon to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Add a new coupon",
                "parameters": [
                    {
                        "description": "Coupon data",
                        "name": "coupon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CouponData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coupon added successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/coupon/applicable": {
            "post": {
                "description": "Returns coupons applicable to the provided cart items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coupons"
                ],
                "summary": "Get applicable coupons",
                "parameters": [
                    {
                        "description": "Cart items",
                        "name": "cart",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.OrderInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of applicable coupons",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/main.ApplicableCoupon"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/coupon/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update coupon details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update an existing coupon",
                "parameters": [
                    {
                        "description": "Coupon update data",
                        "name": "coupon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.UpdateCouponRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/coupon/validate": {
            "post": {
                "description": "Check if a coupon is valid for the given order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coupons"
                ],
                "summary": "Validate a coupon",
                "parameters": [
                    {
                        "description": "Validation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ValidateCoupon"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Validation result",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ApplicableCoupon": {
            "type": "object",
            "properties": {
                "coupon_code": {
                    "type": "string"
                },
                "discount_value": {
                    "type": "number"
                }
            }
        },
        "main.CouponData": {
            "type": "object",
            "required": [
                "coupon_code",
                "discount_Target",
                "discount_type",
                "expiry_date",
                "usage_type",
                "valid_from",
                "valid_until"
            ],
            "properties": {
                "applicable_categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "applicable_medicine_id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "coupon_code": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "discount_Target": {
                    "type": "string",
                    "enum": [
                        "inventory",
                        "charges",
                        "inventory_and_charges"
                    ]
                },
                "discount_type": {
                    "type": "string",
                    "enum": [
                        "flat",
                        "percentage"
                    ]
                },
                "discount_value": {
                    "type": "number"
                },
                "expiry_date": {
                    "type": "string"
                },
                "max_usage_per_user": {
                    "type": "integer"
                },
                "min_order_value": {
                    "type": "number",
                    "minimum": 0
                },
                "terms_and_conditions": {
                    "type": "string"
                },
                "usage_type": {
                    "type": "string",
                    "enum": [
                        "one_time",
                        "multi_use",
                        "time_based"
                    ]
                },
                "valid_from": {
                    "type": "string"
                },
                "valid_until": {
                    "type": "string"
                }
            }
        },
        "main.Medicine": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "main.OrderInput": {
            "type": "object",
            "properties": {
                "cart_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Medicine"
                    }
                },
                "order_total": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "main.UpdateCouponRequest": {
            "type": "object",
            "properties": {
                "coupon_code": {
                    "type": "string"
                },
                "discount_type": {
                    "type": "string"
                },
                "discount_value": {
                    "type": "number"
                },
                "max_usage_per_user": {
                    "type": "integer"
                }
            }
        },
        "main.ValidateCoupon": {
            "type": "object",
            "properties": {
                "cart_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Medicine"
                    }
                },
                "coupon_code": {
                    "type": "string"
                },
                "order_total": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}